# Performance Optimization for computing id hashes.