--- a/src/Internal/IdHashing/ComplexIdHashing.php
+++ b/src/Internal/IdHashing/ComplexIdHashing.php
@@ -11,6 +11,9 @@ use function implode;
 
 class ComplexIdHashing implements IdHashing
 {
+    /**
+     * @param mixed[] $identifier
+     */
     public function getIdHashByIdentifier(array $identifier): string
     {
         return implode(