--- a/src/Internal/IdHashing/SingleElementIdHashing.php
+++ b/src/Internal/IdHashing/SingleElementIdHashing.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Doctrine\ORM\Internal\IdHashing;
+
+use function array_key_first;
+
+class SingleElementIdHashing implements IdHashing
+{
+    public function getIdHashByIdentifier(array $identifier): string
+    {
+        $key = array_key_first($identifier);
+        if (! isset($identifier[$key])) {
+            return ''; // WHY?
+        }
+
+        return $identifier[$key];
+    }
+}