--- a/src/Internal/IdHashing/SingleElementIdHashing.php
+++ b/src/Internal/IdHashing/SingleElementIdHashing.php
@@ -10,11 +10,6 @@ class SingleElementIdHashing implements IdHashing
 {
     public function getIdHashByIdentifier(array $identifier): string
     {
-        $key = array_key_first($identifier);
-        if (! isset($identifier[$key])) {
-            return ''; // WHY?
-        }
-
-        return $identifier[$key];
+        return implode(' ', $identifier);
     }
 }