--- a/src/Internal/IdHashing/SingleElementIdHashing.php
+++ b/src/Internal/IdHashing/SingleElementIdHashing.php
@@ -4,10 +4,13 @@ declare(strict_types=1);
 
 namespace Doctrine\ORM\Internal\IdHashing;
 
-use function array_key_first;
+use function implode;
 
 class SingleElementIdHashing implements IdHashing
 {
+    /**
+     * @param mixed[] $identifier
+     */
     public function getIdHashByIdentifier(array $identifier): string
     {
         return implode(' ', $identifier);