--- a/tests/Tests/ORM/Functional/Ticket/GH7869Test.php
+++ b/tests/Tests/ORM/Functional/Ticket/GH7869Test.php
@@ -55,7 +55,7 @@ class GH7869Test extends OrmTestCase
         $uow->clear();
         $uow->triggerEagerLoads();
 
-        self::assertSame(4, $em->getClassMetadataCalls);
+        self::assertSame(6, $em->getClassMetadataCalls);
     }
 }