In 09b4a75ed3ec3d67f805f6ccc22399997c31f022 we introduced an API `UnitOfWork::getIdHashByIdentifier` that added a significant amount of function calls over the previous inline `implode` to support backed enums as identifiers.

This PR moves ID hashing into a strategy pattern, moving the object to `ClassMetadata` instances of each entity, getting the previous performance almost back (+ the dynamic call). 

I hope this will significantly improve performance of the 80% identifier that is a simple, single numeric identifier. 

Here are the results from two traces compaed with Tideways using the lazy ghost performance project discussed in #11087 that runs `EntityManager::getReference`, hitting the ID hashing code significantly for 100.000 objects.

| Old Call  | Old Time | New Call | New Time |
| ------------- | ------------- | ------------- | ------------- |
| `UnitOfWork::getIdHashByIdentifier` | 46.75 ms  | `SingleElementIdHashing::getIdHashByIdentifier` | 15.3 ms |
| `UnitOfWork::tryGetById` | 33.01 ms | `UnitOfWork::tryGetByIdWithClass` | 16.1 ms |

<img width="640" alt="Bildschirmfoto 2024-02-28 um 22 22 28" src="https://github.com/doctrine/orm/assets/26936/c315b9f2-e3d9-4f0e-9ef5-30e210f70268">

While not life-changing, this hit accrues, especially in test-suites working with fixtures.